# openapi.yaml
openapi: "3.0.3"
info:
    version: 1.0.0
    title: ETAPI
    description: External Trilium API
    contact:
        name: zadam
        email: zadam.apps@gmail.com
        url: https://github.com/zadam/trilium
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
    - url: http://localhost:37740/etapi
    - url: http://localhost:8080/etapi
security:
    - EtapiTokenAuth: []
    - EtapiBasicAuth: []
paths:
    /create-note:
        post:
            description: Create a note and place it into the note tree
            operationId: createNote
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateNoteDef"
            responses:
                "201":
                    description: note created
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/NoteWithBranch"
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
...
    /notes/{noteId}:
        parameters:
            - name: noteId
              in: path
              required: true
              schema:
                  $ref: "#/components/schemas/EntityId"
        get:
            description: Returns a note identified by its ID
            operationId: getNoteById
            responses:
                "200":
                    description: note response
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Note"
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
        patch:
            description: patch a note identified by the noteId with changes in the body
            operationId: patchNoteById
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Note"
            responses:
                "200":
                    description: note updated
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Note"
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
        delete:
            description: deletes a single note based on the noteId supplied
            operationId: deleteNoteById
            responses:
                "204":
                    description: note deleted
                default:
                    description: unexpected error
                    content:
                        application/json; charset=utf-8:
                            schema:
                                $ref: "#/components/schemas/Error"
    /notes/{noteId}/content:
        parameters:
            - name: noteId
              in: path
              required: true
              schema:
                  $ref: "#/components/schemas/EntityId"
        get:
            description: Returns note content identified by its ID
            operationId: getNoteContent
            responses:
                "200":
                    description: note content response
                    content:
                        text/html:
                            schema:
                                type: string
        put:
            description: Updates note content identified by its ID
            operationId: putNoteContentById
            requestBody:
                description: html content of note
                required: true
                content:
                    text/plain:
                        schema:
                            type: string
            responses:
                "204":
                    description: note content updated
...
components:
    securitySchemes:
        EtapiTokenAuth:
            type: apiKey
            in: header
            name: Authorization
        EtapiBasicAuth:
            type: http
            scheme: basic
            description: >
                Basic Auth where username is arbitrary string (e.g. "trilium", not checked), 
                username is the ETAPI token. 
                To emphasize, do not use Trilium password here (won't work), only the generated
                ETAPI token (from Options -> ETAPI)
    schemas:
        CreateNoteDef:
            type: object
            required:
                - parentNoteId
                - title
                - type
                - content
            properties:
                parentNoteId:
                    $ref: "#/components/schemas/EntityId"
                    description: Note ID of the parent note in the tree
                title:
                    type: string
                type:
                    type: string
                    enum:
                        - text
                        - code
                        - file
                        - image
                        - search
                        - book
                        - relationMap
                        - render
                mime:
                    type: string
                    description: this needs to be specified only for note types 'code', 'file', 'image'.
                    example: application/json
                content:
                    type: string
                notePosition:
                    type: integer
                    description: >
                        Position of the note in the parent. Normal ordering is 10, 20, 30 ... 
                        So if you want to create a note on the first position, use e.g. 5, for second position 15, for last e.g. 1000000
                prefix:
                    type: string
                    description: >
                        Prefix is branch (placement) specific title prefix for the note. 
                        Let's say you have your note placed into two different places in the tree, 
                        but you want to change the title a bit in one of the placements. For this you can use prefix.
                isExpanded:
                    type: boolean
                    description: true if this note (as a folder) should appear expanded
                noteId:
                    $ref: "#/components/schemas/EntityId"
                    description: DON'T specify unless you want to force a specific noteId
                branchId:
                    $ref: "#/components/schemas/EntityId"
                    description: DON'T specify unless you want to force a specific branchId
                dateCreated:
                    $ref: "#/components/schemas/LocalDateTime"
                    description: Local timestap of the note creation. Specify only if you want to override the default (current datetime in the current timezone/offset).
                utcDateCreated:
                    $ref: "#/components/schemas/UtcDateTime"
                    description: UTC timestap of the note creation. Specify only if you want to override the default (current datetime).
        Note:
            type: object
            properties:
                noteId:
                    $ref: "#/components/schemas/EntityId"
                    readOnly: true
                title:
                    type: string
                type:
                    type: string
                    enum: [text, code, render, file, image, search, relationMap, book, noteMap, mermaid, webView, shortcut, doc, contentWidget, launcher]
                mime:
                    type: string
                isProtected:
                    type: boolean
                    readOnly: true
                blobId:
                    type: string
                    description: ID of the blob object which effectively serves as a content hash
                attributes:
                    $ref: "#/components/schemas/AttributeList"
                    readOnly: true
                parentNoteIds:
                    $ref: "#/components/schemas/EntityIdList"
                    readOnly: true
                childNoteIds:
                    $ref: "#/components/schemas/EntityIdList"
                    readOnly: true
                parentBranchIds:
                    $ref: "#/components/schemas/EntityIdList"
                    readOnly: true
                childBranchIds:
                    $ref: "#/components/schemas/EntityIdList"
                    readOnly: true
                dateCreated:
                    $ref: "#/components/schemas/LocalDateTime"
                dateModified:
                    $ref: "#/components/schemas/LocalDateTime"
                    readOnly: true
                utcDateCreated:
                    $ref: "#/components/schemas/UtcDateTime"
                utcDateModified:
                    $ref: "#/components/schemas/UtcDateTime"
                    readOnly: true
        Branch:
            type: object
            description: Branch places the note into the tree, it represents the relationship between a parent note and child note
            properties:
                branchId:
                    $ref: "#/components/schemas/EntityId"
                noteId:
                    $ref: "#/components/schemas/EntityId"
                    readOnly: true
                    description: identifies the child note
                parentNoteId:
                    $ref: "#/components/schemas/EntityId"
                    readOnly: true
                    description: identifies the parent note
                prefix:
                    type: string
                notePosition:
                    type: integer
                    format: int32
                isExpanded:
                    type: boolean
                utcDateModified:
                    $ref: "#/components/schemas/UtcDateTime"
                    readOnly: true
        NoteWithBranch:
            type: object
            properties:
                note:
                    $ref: "#/components/schemas/Note"
                branch:
                    $ref: "#/components/schemas/Branch"
...
        Attribute:
            type: object
            description: Attribute (Label, Relation) is a key-value record attached to a note.
            properties:
                attributeId:
                    $ref: "#/components/schemas/EntityId"
                noteId:
                    $ref: "#/components/schemas/EntityId"
                    readOnly: true
                    description: identifies the child note
                type:
                    type: string
                    enum: [label, relation]
                name:
                    type: string
                    pattern: '^[^\s]+'
                    example: shareCss
                value:
                    type: string
                position:
                    type: integer
                    format: int32
                isInheritable:
                    type: boolean
                utcDateModified:
                    $ref: "#/components/schemas/UtcDateTime"
                    readOnly: true
        AttributeList:
            type: array
            items:
                $ref: "#/components/schemas/Attribute"
...
        EntityId:
            type: string
            pattern: "[a-zA-Z0-9_]{4,32}"
            example: evnnmvHTCgIn
...
        EntityIdList:
            type: array
            items:
                $ref: "#/components/schemas/EntityId"
        LocalDateTime:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}[\+\-][0-9]{4}'
            example: 2021-12-31 20:18:11.930+0100
        UtcDateTime:
            type: string
            pattern: '[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}Z'
            example: 2021-12-31 19:18:11.930Z
...
        Error:
            type: object
            required:
                - status
                - code
                - message
            properties:
                status:
                    type: integer
                    format: int32
                    description: HTTP status, identical to the one given in HTTP response
                    example: 400
                code:
                    type: string
                    description: stable string constant
                    example: NOTE_IS_PROTECTED
                message:
                    type: string
                    description: Human readable error, potentially with more details,
                    example: Note 'evnnmvHTCgIn' is protected and cannot be modified through ETAPI